Add __new__ to the OneHotEncoder class
