+++ src/skprometheus/preprocessing.py
@@ -9,14 +9,18 @@ class OneHotEncoder(preprocessing.OneHotEncoder):
     """
     OneHotEncoder that adds metrics to the prometheus metric registry.
     """
-    @wraps(preprocessing.OneHotEncoder.__init__, assigned=["__signature__"])
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
+
+    def __new__(cls, *args, **kwargs):
         MetricRegistry.add_counter(
             "model_categorical",
             "Counts category occurrence for each categorical feature.",
             additional_labels=("feature", "category"),
         )
+        return super(OneHotEncoder, cls).__new__(cls)
+    @wraps(preprocessing.OneHotEncoder.__init__, assigned=["__signature__"])
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
 
     def transform(self, X):
         """
